+++ spec/support/examples_message.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+shared_examples 'message is' do |type|
+  it { expect(message).to be_kind_of(type) }
+end